+++ database.js
@@ -26,6 +26,7 @@ const designer  = require('./src/db/designer');
 const importdock = require('./src/db/importdock');
 const versions  = require('./src/db/versions');
 const migrate   = require('./src/db/migrate_v3');
+const sync      = require('./src/db/sync');
 
 module.exports = {
   ...core,
@@ -55,4 +56,5 @@ module.exports = {
   ...importdock,
   ...versions,
   ...migrate,
+  ...sync,
 };